requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
openai
PyMuPDF
python-docx
chardet
pandas
Pillow